rename(Bets, Money): 패키지 이동
