+++ src/main/java/blackjack/BlackJackGame.java
@@ -3,8 +3,8 @@ package blackjack;
 import blackjack.dto.NameCards;
 import blackjack.dto.NameCardsScore;
 import blackjack.dto.NameProfit;
-import blackjack.model.Bets;
-import blackjack.model.Money;
+import blackjack.model.money.Bets;
+import blackjack.model.money.Money;
 import blackjack.model.deck.Deck;
 import blackjack.model.participant.Dealer;
 import blackjack.model.participant.Player;
