+++ src/main/java/blackjack/model/money/Money.java
@@ -1,4 +1,4 @@
-package blackjack.model;
+package blackjack.model.money;
 
 import java.util.Objects;
