+++ src/test/java/blackjack/model/money/MoneyTest.java
@@ -1,12 +1,10 @@
-package blackjack.model;
+package blackjack.model.money;
 
 import static org.assertj.core.api.Assertions.assertThat;
-import static org.assertj.core.api.Assertions.assertThatIllegalArgumentException;
 
+import blackjack.model.money.Money;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.junit.jupiter.params.ParameterizedTest;
-import org.junit.jupiter.params.provider.CsvSource;
 
 public class MoneyTest {
